Address PR review: split ForecasterModule into an abstract base + concrete deterministic/probabilistic modules
Introduce BaseForecasterModule (abstract) under models/forecasters/ holding
shared plumbing (training_step, common_step, batch standardization,
checkpoint compatibility, plotting/aggregation helpers), with
validation_step, test_step and on_test_epoch_end left abstract since they
differ meaningfully between evaluation modes. Rename ForecasterModule to
DeterministicForecasterModule and move it, alongside
ProbabilisticForecasterModule, into forecasters/ as siblings implementing
the shared contract, rather than one subclassing the other.
