+++ tests/test_prediction_model_classes.py
@@ -7,7 +7,11 @@ import torch
 
 # First-party
 from neural_lam import config as nlconfig
-from neural_lam.models import ARForecaster, ForecasterModule, StepPredictor
+from neural_lam.models import (
+    ARForecaster,
+    DeterministicForecasterModule,
+    StepPredictor,
+)
 from tests.conftest import init_datastore_example
 from tests.dummy_datastore import DummyDatastore
 
@@ -81,7 +85,7 @@ def test_forecaster_module_checkpoint(tmp_path):
     )
 
     # Build predictor and forecaster externally, then inject into
-    # ForecasterModule
+    # DeterministicForecasterModule
     # First-party
     from neural_lam.models import MODELS
 
@@ -99,7 +103,7 @@ def test_forecaster_module_checkpoint(tmp_path):
     )
     forecaster = ARForecaster(predictor, datastore, config=config, loss="mse")
 
-    model = ForecasterModule(
+    model = DeterministicForecasterModule(
         forecaster=forecaster,
         config=config,
         datastore=datastore,
@@ -137,7 +141,7 @@ def test_forecaster_module_checkpoint(tmp_path):
     )
 
     # Load from checkpoint
-    loaded_model = ForecasterModule.load_from_checkpoint(
+    loaded_model = DeterministicForecasterModule.load_from_checkpoint(
         ckpt_path,
         datastore=datastore,
         forecaster=load_forecaster,
@@ -195,7 +199,7 @@ def test_forecaster_module_old_checkpoint(tmp_path):
         output_std=False,
     )
     # Use distinctive non-default values so we can detect silent fallback
-    # to ForecasterModule's defaults during load.
+    # to DeterministicForecasterModule's defaults during load.
     saved_loss = "mse"
     saved_lr = 0.123
     saved_create_gif = True
@@ -206,7 +210,7 @@ def test_forecaster_module_old_checkpoint(tmp_path):
         predictor, datastore, config=config, loss=saved_loss
     )
 
-    model = ForecasterModule(
+    model = DeterministicForecasterModule(
         forecaster=forecaster,
         config=config,
         datastore=datastore,
@@ -276,7 +280,7 @@ def test_forecaster_module_old_checkpoint(tmp_path):
     )
 
     # Load from hacked old checkpoint
-    loaded_model = ForecasterModule.load_from_checkpoint(
+    loaded_model = DeterministicForecasterModule.load_from_checkpoint(
         ckpt_path,
         datastore=datastore,
         forecaster=load_forecaster,
@@ -287,7 +291,8 @@ def test_forecaster_module_old_checkpoint(tmp_path):
     assert loaded_model.forecaster.predictor.__class__.__name__ == "GraphLAM"
 
     # Hyperparameters nested in the legacy 'args' namespace must round-trip
-    # rather than silently falling back to ForecasterModule defaults.
+    # rather than silently falling back to DeterministicForecasterModule
+    # defaults.
     assert loaded_model.hparams.lr == saved_lr
     assert loaded_model.hparams.val_steps_to_log == saved_val_steps
     assert loaded_model.create_gif is saved_create_gif
